Bao/back end

+++ back-end/back-end/Service/Implement/PostService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using back_end.Common;
 using back_end.DTO;
 using back_end.DTO.UserDTOModel;
 using back_end.Enity;
@@ -8,6 +9,7 @@ using back_end.Respositories.Implement;
 using back_end.Respositories.Interface;
 using back_end.Service.Interface;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace back_end.Service.Implement
 {
@@ -17,6 +19,9 @@ namespace back_end.Service.Implement
         private IPostTechnologyRespository _postTechnologyRespository;
         private ITechnologyRespository _technologyRespository;
         private IEmployeeRespository _employeeRespository;
+        private IHttpContextAccessor _contextAccessor;
+        private IUserRespository _userRespository;
+        private IUserService _userService;
         private IMapper _mapper;
 
         public PostService(
@@ -24,12 +29,18 @@ namespace back_end.Service.Implement
             IPostTechnologyRespository postTechnologyRespository,
             ITechnologyRespository technologyRespository,  
             IEmployeeRespository employeeRespository,
+            IHttpContextAccessor contextAccessor,
+            IUserRespository userRespository,
+            IUserService userService,
             IMapper mapper)
         {
             _postRespository = postRespository;
             _postTechnologyRespository = postTechnologyRespository;
             _technologyRespository = technologyRespository;
             _employeeRespository = employeeRespository;
+            _contextAccessor = contextAccessor;
+            _userRespository = userRespository;
+            _userService = userService;
             _mapper = mapper;
         }
 
@@ -38,43 +49,46 @@ namespace back_end.Service.Implement
             var result = new AppResponse<PostDTO>();
             try
             {
+
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Can't found user from token");
+
                 // ------------ check employee and  permission -------------- //
                 var employee = await _employeeRespository
-                    .FindBy(em => !em.IsDelete && em.Id == post.EmployeeId)
-                    .Include(em=>em.User)
-                    .ThenInclude(u => u.UserRoles)
-                    .ThenInclude(ur=>ur.Role)
+                    .FindBy(em => !em.IsDelete && em.UserId == user.Id)
                     .FirstOrDefaultAsync();
-                if(
-                    employee is null 
-                    || employee.User.UserRoles.Where(ur => ur.Role.Name.Equals("business")).FirstOrDefault() is null
-                    || employee.Type != EmployeeType.OfficalEmployee 
-                ) return result.BuilderError("You don't have permission");
+                if( employee is null || employee.Type != EmployeeType.OfficalEmployee) 
+                    return result.BuilderError("You don't have permission");
 
 
                 // -------- saving new employee ---------------//
                 Post newPost = _mapper.Map<Post>(post);
                 newPost.InitialEnity();
+                newPost.EmployeeId = employee.Id;
                 await _postRespository.Insert(newPost);
 
                 // -------- saving post of technologies ------- //
                 List<PostTechnology> postTechnologyList = new List<PostTechnology>();
-                var techList = post.Technologies.Select(t => _mapper.Map<Technology>(t)).ToList();
+                var techList = post.Technologies?.Select(t => _mapper.Map<Technology>(t)).ToList();
                 post.Id = newPost.Id;
                 post.Technologies = [];
-                foreach (var technology in techList)
+                if(techList != null)
                 {
-                    var tech = await _technologyRespository.FindBy(t => t.Id == technology.Id && !t.IsDelete).FirstOrDefaultAsync();
-                    if (tech != null){
-                        var postTechnology = new PostTechnology();
-                        postTechnology.InitialEnity();
-                        postTechnology.PostId = newPost.Id;
-                        postTechnology.TechnologyId = tech.Id;
-                        post.Technologies.Add(_mapper.Map<TechnologyDTO>(tech));
-                        postTechnologyList.Add(postTechnology);
+                    foreach (var technology in techList)
+                    {
+                        var tech = await _technologyRespository.FindBy(t => t.Id == technology.Id && !t.IsDelete).FirstOrDefaultAsync();
+                        if (tech != null){
+                            var postTechnology = new PostTechnology();
+                            postTechnology.InitialEnity();
+                            postTechnology.PostId = newPost.Id;
+                            postTechnology.TechnologyId = tech.Id;
+                            post.Technologies.Add(_mapper.Map<TechnologyDTO>(tech));
+                            postTechnologyList.Add(postTechnology);
+                        }
                     }
+                    await _postTechnologyRespository.Insert(postTechnologyList);
                 }
-                await _postTechnologyRespository.Insert(postTechnologyList);
 
                 return result.BuilderResult(post,"Success");
             }
@@ -90,9 +104,15 @@ namespace back_end.Service.Implement
             var result = new AppResponse<bool>();
             try
             {
-                var post = await _postRespository.FindBy(p => p.Id == postId).FirstOrDefaultAsync();
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Not found user from token");
+
+                var post = await _postRespository.FindBy(p => p.Id == postId).Include(p=>p.Employee).FirstOrDefaultAsync();
                 if (post is null) return result.BuilderError("Not found post");
 
+                if (post.Employee.UserId != user.Id && !_userService.checkRole())
+                    return result.BuilderError("You are not Author of this post");
                 post.DeleteEnity();
                 await _postRespository.Update(post);
 
@@ -145,12 +165,19 @@ namespace back_end.Service.Implement
             var result = new AppResponse<PostDTO>();
             try
             {
-                var data = await _postRespository
+                Post? data = await _postRespository
                     .FindBy(p => !p.IsDelete && p.Id == post.Id)
                     .Include(p => p.Employee)
                     .Include(p=>p.PostTechnologies).ThenInclude(pt => pt.Technology)
                     .FirstOrDefaultAsync();
                 if (data is null) return result.BuilderError("Not found post");
+
+                User user = await _userService.GetUserFromToken();
+                if (user is null)
+                    return result.BuilderError("Can't found user from token");
+
+                if (user.Id != data.Employee.UserId && !_userService.checkRole())
+                    return result.BuilderError("You are not Author of this Post");
 
                 // ----------- Update information post ------------------- //
